Add preprocess example.
This is the first example that has made the difference in behaviour
between the C code (changing pointers to be null when destroying
object) and the Ruby code (just pretending) has become an issue.
I knew this would be an issue from the beginning but I haven't
yet worked out the best Ruby-ish way to do it.